PatientApp/requirements.txt
appdirs==1.4.3
asgiref==3.3.4
astroid==2.4.1
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
colorama==0.4.3
cryptography==3.4.7
dataclasses==0.6
dialogflow==1.1.0
distlib==0.3.0
dj-database-url==0.5.0
Django==2.2.8
django-filter==2.4.0
django-google-api==1.0.1
django-heroku==0.3.1
django-jsonfield==1.4.1
djangorestframework==3.12.4
djangotoolbox==1.8.0
djongo==1.2.31
fasteners==0.16.2
filelock==3.0.12
future==0.18.2
google-api-core==1.30.0
google-api-python-client==2.8.0
google-auth==1.30.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
google-cloud==0.34.0
google-cloud-dialogflow==2.4.0
google-oauth==1.0.1
google-oauth2l==1.0.2
googleapis-common-protos==1.53.0
grpcio==1.38.0
httplib2==0.19.1
idna==2.10
isort==4.3.21
joblib==1.0.1
lazy-object-proxy==1.4.3
Markdown==3.3.4
mccabe==0.6.1
nltk==3.6.2
numpy==1.20.3
oauth2client==4.1.3
oauthlib==3.1.1
packaging==20.9
pandas==1.2.4
plotly==4.14.3
proto-plus==1.18.1
protobuf==3.17.3
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pylint==2.5.2
pymongo==3.11.4
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2020.1
rake-nltk==1.0.4
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.7.2
six==1.14.0
sqlparse==0.2.4
stopwords==1.0.0
style==1.1.0
toml==0.10.0
tqdm==4.61.1
update==0.0.1
uritemplate==3.0.1
urllib3==1.26.5
virtualenv==20.0.20
virtualenvwrapper-win==1.2.6
whitenoise==5.2.0
wrapt==1.12.1
